+++ test/lib/swagger/formatter_test.exs
@@ -3,7 +3,7 @@ defmodule Xcribe.Swagger.FormatterTest do
 
   alias Xcribe.Request
   alias Xcribe.Support.Information, as: ExampleInformation
-  alias Xcribe.Support.Samples.SwaggerFormater, as: Samples
+  alias Xcribe.Support.Samples.SwaggerFormater.PathItemObject, as: Samples
   alias Xcribe.Swagger.Formatter
 
   describe "raw_openapi_object/0" do
@@ -572,36 +572,74 @@ defmodule Xcribe.Swagger.FormatterTest do
   end
 
   describe "merge_parameter_object_lists/2" do
-    test "keep uniq names and " do
-      base_list = [
+    test "keep uniq names and order by name" do
+      parameters = [
         %{name: "id", in: "path", required: true, schema: %{type: "string"}, example: 6},
         %{name: "id", in: "header", schema: %{type: "string"}, example: "8"}
       ]
 
-      new_list = [
+      new_params = [
         %{name: "id", in: "path", required: true, schema: %{type: "string"}, example: 9},
-        %{name: "id", in: "query", schema: %{type: "string"}, example: "9090"},
-        %{name: "alias", in: "query", schema: %{type: "string"}, example: "jon"}
+        %{name: "alias", in: "query", schema: %{type: "string"}, example: "jon"},
+        %{name: "id", in: "query", schema: %{type: "string"}, example: "9090"}
       ]
 
       expected = [
         %{name: "alias", in: "query", schema: %{type: "string"}, example: "jon"},
-        %{name: "id", in: "query", schema: %{type: "string"}, example: "9090"},
+        %{name: "id", in: "header", schema: %{type: "string"}, example: "8"},
         %{name: "id", in: "path", required: true, schema: %{type: "string"}, example: 6},
-        %{name: "id", in: "header", schema: %{type: "string"}, example: "8"}
+        %{name: "id", in: "query", schema: %{type: "string"}, example: "9090"}
       ]
 
-      assert Formatter.merge_parameter_object_lists(base_list, new_list) == expected
+      assert Formatter.merge_parameter_object_lists(parameters, new_params) == expected
+    end
+
+    test "overwrite params with new one" do
+      parameters = [
+        %{
+          name: "id",
+          in: "path",
+          required: true,
+          schema: %{type: "string"},
+          example: "invalid-id"
+        }
+      ]
+
+      new_params = [
+        %{name: "id", in: "path", required: true, schema: %{type: "string"}, example: "valid-id"}
+      ]
+
+      assert Formatter.merge_parameter_object_lists(parameters, new_params, :overwrite) ==
+               new_params
     end
   end
 
   describe "merge_path_item_objects/3" do
-    test "merge 2 object with diferente responses" do
-      object_one = Samples.path_item_object_without_request_body()
-      object_two = Samples.path_item_object_with_request_body()
-      expected = Samples.expected_path_objects_merge()
+    test "merge objects with diferente responses and choose correct examples" do
+      not_found = Samples.not_found_without_req_body()
+      bad_request = Samples.bad_request_with_req_body()
+      success = Samples.success_with_req_body()
 
-      assert Formatter.merge_path_item_objects(object_one, object_two, "put") == expected
+      first_sequence =
+        not_found
+        |> Formatter.merge_path_item_objects(bad_request, "put")
+        |> Formatter.merge_path_item_objects(success, "put")
+
+      second_sequence =
+        success
+        |> Formatter.merge_path_item_objects(not_found, "put")
+        |> Formatter.merge_path_item_objects(bad_request, "put")
+
+      third_sequence =
+        bad_request
+        |> Formatter.merge_path_item_objects(success, "put")
+        |> Formatter.merge_path_item_objects(not_found, "put")
+
+      expected = Samples.all_merged()
+
+      assert first_sequence == expected
+      assert second_sequence == expected
+      assert third_sequence == expected
     end
   end
 end